Add files via upload
Continuing to work on the transect summary figures. Added by season summaries to help think about what we are going to do given most of the historic sampling has occurred during "summer" and less during "winter".  So is an annual sample a summer + winter sampling event?  Or just one or the other? Given the variation between summer and winter we had chosen to sample winter, but, there just aren't that many winter samples in the database to work with.